requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
Werkzeug==3.1.3
